Move Liveness up one level
